Minimize unnecessary additional files
