--- src/utils/generateEntityKey.ts
@@ -1,3 +1,0 @@
-export function generateEntityKey(type: string, id: string | number) {
-  return `${type}_${id}`;
-}
